feat(connection): default to main
