+++ packages/wallet/ui/src/contexts/Application.jsx
@@ -9,15 +9,13 @@ export const ConnectionStatus = {
   Error: 'error',
 };
 
-export const useApplicationContext = () => useContext(ApplicationContext);
-
 // Higher-order component wrapper for mapping context to props. This allows
 // components to use `memo` and avoid rerendering when unrelated context
 // changes.
 export const withApplicationContext = (Component, mapContextToProps) => {
   const MemoizedComponent = memo(Component);
   return ({ ...props }) => {
-    const context = mapContextToProps(useApplicationContext());
+    const context = mapContextToProps(useContext(ApplicationContext));
 
     return <MemoizedComponent {...context} {...props} />;
   };
